fix(monetary): Fase 1 - corregir residuos críticos de float (ADR-018)
CORRECCIONES APLICADAS:

1. TipPayout Entity:
   - Cambiado: 'amount' => 'decimal:2'
   - A: 'amount' => 'integer'
   - Razón: Las propinas son montos en CLP, deben ser integer

2. CashSession Entity:
   - Línea 143: (float) $tipPayoutQuery->sum('amount')
   - Cambiado a: (int) $tipPayoutQuery->sum('amount')
   - Razón: Suma de propinas debe ser integer

3. RefundService:
   - Reemplazados todos los (float) por (int)
   - Montos de reembolso ahora son integer
   - Razón: Los reembolsos son montos en CLP

IMPACTO:
✅ TipPayout: modelo monetario correcto
✅ CashSession: cálculos de cierre de caja correctos
✅ RefundService: reembolsos en integer end-to-end

ADRs respetados:
✅ ADR-018: Integridad financiera (sin floats en montos)
✅ ADR-011: Modelo monetario chileno (CLP entero)

PRÓXIMAS FASES:
- Fase 2: Agregar casts explícitos a entidades (Order, Bill, Payment)
- Fase 3: Migrar servicios no-críticos (Kitchen, Inventory, Fiscal)

+++ app/Modules/Payments/Domain/Entities/CashSession.php
@@ -140,7 +140,7 @@ class CashSession extends Model
             $tipPayoutQuery->where('payment_method', 'cash');
         }
         
-        $tipsPaidOut = (float) $tipPayoutQuery->sum('amount');
+        $tipsPaidOut = (int) $tipPayoutQuery->sum('amount');
         
         // Calcular impacto de movimientos (si se solicita)
         $movementsImpact = 0;
